* Correctly register MIME type. * Don't try to associate with the MIME type for now, because it doesn't   really work yet. * Make the weight entry box allow floats. -Samuel

+++ src/day.py
@@ -25,7 +25,7 @@ class Day:
         return self.date
 
     def set_weight(self, weight):
-        self.weight = int(weight)
+        self.weight = float(weight)
         
     def get_weight(self):
         return self.weight
